+++ component/receiver_test.go
@@ -46,31 +46,9 @@ func TestNewReceiverFactory_WithOptions(t *testing.T) {
 	factory := NewReceiverFactory(
 		typeStr,
 		func() config.Receiver { return &defaultCfg },
-		WithTracesReceiver(createTracesReceiver),
-		WithMetricsReceiver(createMetricsReceiver),
-		WithLogsReceiver(createLogsReceiver))
-	assert.EqualValues(t, typeStr, factory.Type())
-	assert.EqualValues(t, &defaultCfg, factory.CreateDefaultConfig())
-
-	_, err := factory.CreateTracesReceiver(context.Background(), ReceiverCreateSettings{}, &defaultCfg, nil)
-	assert.NoError(t, err)
-
-	_, err = factory.CreateMetricsReceiver(context.Background(), ReceiverCreateSettings{}, &defaultCfg, nil)
-	assert.NoError(t, err)
-
-	_, err = factory.CreateLogsReceiver(context.Background(), ReceiverCreateSettings{}, &defaultCfg, nil)
-	assert.NoError(t, err)
-}
-
-func TestNewReceiverFactory_WithStabilityLevels(t *testing.T) {
-	const typeStr = "test"
-	defaultCfg := config.NewReceiverSettings(config.NewComponentID(typeStr))
-	factory := NewReceiverFactory(
-		typeStr,
-		func() config.Receiver { return &defaultCfg },
-		WithTracesReceiverAndStabilityLevel(createTracesReceiver, StabilityLevelDeprecated),
-		WithMetricsReceiverAndStabilityLevel(createMetricsReceiver, StabilityLevelAlpha),
-		WithLogsReceiverAndStabilityLevel(createLogsReceiver, StabilityLevelStable))
+		WithTracesReceiver(createTracesReceiver, StabilityLevelDeprecated),
+		WithMetricsReceiver(createMetricsReceiver, StabilityLevelAlpha),
+		WithLogsReceiver(createLogsReceiver, StabilityLevelStable))
 	assert.EqualValues(t, typeStr, factory.Type())
 	assert.EqualValues(t, &defaultCfg, factory.CreateDefaultConfig())