+++ CHANGELOG.md
@@ -7,9 +7,30 @@
 - Remove deprecated funcs/types from service related to `Config` (#5755)
 - Change`confighttp.ToClient` to accept a `component.Host` (#5737)
 - Remove deprecated funcs from pdata related to mutable slices (#5754)
+- Change the following deprecated component functions to ensure a stability level is set:
+  - `component.WithTracesExporter`
+  - `component.WithMetricsExporter`
+  - `component.WithLogsExporter`
+  - `component.WithTracesReceiver`
+  - `component.WithMetricsReceiver`
+  - `component.WithLogsReceiver`
+  - `component.WithTracesProcessor`
+  - `component.WithMetricsProcessor`
+  - `component.WithLogsProcessor`
 
 ### 🚩 Deprecations 🚩
 
+- Deprecate the following component functions added to ensure a stability level is set:
+  - `component.WithTracesExporterAndStabilityLevel` -> `component.WithTracesExporter`
+  - `component.WithMetricsExporterAndStabilityLevel` -> `component.WithMetricsExporter`
+  - `component.WithLogsExporterAndStabilityLevel` -> `component.WithLogsExporter`
+  - `component.WithTracesReceiverAndStabilityLevel` -> `component.WithTracesReceiver`
+  - `component.WithMetricsReceiverAndStabilityLevel` -> `component.WithMetricsReceiver`
+  - `component.WithLogsReceiverAndStabilityLevel` -> `component.WithLogsReceiver`
+  - `component.WithTracesProcessorAndStabilityLevel` -> `component.WithTracesProcessor`
+  - `component.WithMetricsProcessorAndStabilityLevel` -> `component.WithMetricsProcessor`
+  - `component.WithLogsProcessorAndStabilityLevel` -> `component.WithLogsProcessor`
+  - 
 ### 💡 Enhancements 💡
 
 - `ocb` now exits with an error if it fails to load the build configuration. (#5731)